feat(api): refactorizacion modular de la API REST, enrutamiento y CORS (Closes #3)
