+++ newsradar_api/app/main.py
@@ -1,12 +1,6 @@
-from __future__ import annotations
-
-from datetime import datetime, timezone
-from typing import Dict, List, Optional
-from uuid import uuid4
-
-from fastapi import Depends, FastAPI, HTTPException, Response, status
-from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
-from pydantic import BaseModel, EmailStr, Field, HttpUrl
+from fastapi import FastAPI
+from app.api.router import api_router
+from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI(
     title="NewsRadar API",
@@ -14,869 +8,17 @@ app = FastAPI(
     description="API REST para gestión de usuarios, alertas, notificaciones, fuentes y canales RSS.",
 )
 
-API_PREFIX = "/api/v1"
-security = HTTPBearer(auto_error=False)
-
-
-class Metric(BaseModel):
-    name: str = Field(..., min_length=1, max_length=100)
-    value: float
-
-
-class RoleBase(BaseModel):
-    name: str = Field(..., min_length=1, max_length=100)
-
-
-class RoleCreate(RoleBase):
-    pass
-
-
-class RoleUpdate(BaseModel):
-    name: Optional[str] = Field(None, min_length=1, max_length=100)
-
-
-class Role(RoleBase):
-    id: int
-
-
-class UserBase(BaseModel):
-    email: EmailStr
-    first_name: str = Field(..., min_length=1, max_length=120)
-    last_name: str = Field(..., min_length=1, max_length=120)
-    organization: str = Field(..., min_length=1, max_length=180)
-    role_ids: List[int] = Field(default_factory=list)
-
-
-class UserCreate(UserBase):
-    password: str = Field(..., min_length=6, max_length=128)
-
-
-class UserUpdate(BaseModel):
-    email: Optional[EmailStr] = None
-    first_name: Optional[str] = Field(None, min_length=1, max_length=120)
-    last_name: Optional[str] = Field(None, min_length=1, max_length=120)
-    organization: Optional[str] = Field(None, min_length=1, max_length=180)
-    role_ids: Optional[List[int]] = None
-    password: Optional[str] = Field(None, min_length=6, max_length=128)
-
-
-class User(UserBase):
-    id: int
-
-
-class UserInDB(User):
-    password: str
-
-
-class AlertCategoryItem(BaseModel):
-    code: str = Field(..., min_length=1, max_length=60)
-    label: str = Field(..., min_length=1, max_length=120)
-
-
-class AlertBase(BaseModel):
-    name: str = Field(..., min_length=1, max_length=200)
-    descriptors: List[str] = Field(default_factory=list)
-    categories: List[AlertCategoryItem] = Field(default_factory=list)
-    cron_expression: str = Field(..., min_length=1, max_length=120)
-
-
-class AlertCreate(AlertBase):
-    pass
-
-
-class AlertUpdate(BaseModel):
-    name: Optional[str] = Field(None, min_length=1, max_length=200)
-    descriptors: Optional[List[str]] = None
-    categories: Optional[List[AlertCategoryItem]] = None
-    cron_expression: Optional[str] = Field(None, min_length=1, max_length=120)
-
-
-class Alert(AlertBase):
-    id: int
-    user_id: int
-
-
-class CategoryBase(BaseModel):
-    name: str = Field(..., min_length=1, max_length=120)
-    source: str = Field(default="IPTC", pattern="^IPTC$")
-
-
-class CategoryCreate(CategoryBase):
-    pass
-
-
-class CategoryUpdate(BaseModel):
-    name: Optional[str] = Field(None, min_length=1, max_length=120)
-    source: Optional[str] = Field(None, pattern="^IPTC$")
-
-
-class Category(CategoryBase):
-    id: int
-
-
-class NotificationBase(BaseModel):
-    timestamp: datetime
-    metrics: List[Metric] = Field(default_factory=list)
-
-
-class NotificationCreate(NotificationBase):
-    pass
-
-
-class NotificationUpdate(BaseModel):
-    timestamp: Optional[datetime] = None
-    metrics: Optional[List[Metric]] = None
-
-
-class Notification(NotificationBase):
-    id: int
-    alert_id: int
-
-
-class InformationSourceBase(BaseModel):
-    name: str = Field(..., min_length=1, max_length=120)
-    url: HttpUrl
-
-
-class InformationSourceCreate(InformationSourceBase):
-    pass
-
-
-class InformationSourceUpdate(BaseModel):
-    name: Optional[str] = Field(None, min_length=1, max_length=120)
-    url: Optional[HttpUrl] = None
-
-
-class InformationSource(InformationSourceBase):
-    id: int
-
-
-class RSSChannelBase(BaseModel):
-    url: HttpUrl
-    category_id: int
-
-
-class RSSChannelCreate(RSSChannelBase):
-    pass
-
-
-class RSSChannelUpdate(BaseModel):
-    url: Optional[HttpUrl] = None
-    category_id: Optional[int] = None
-
-
-class RSSChannel(RSSChannelBase):
-    id: int
-    information_source_id: int
-
-
-class StatsBase(BaseModel):
-    metrics: List[Metric] = Field(default_factory=list)
-
-
-class StatsCreate(StatsBase):
-    pass
-
-
-class StatsUpdate(BaseModel):
-    metrics: Optional[List[Metric]] = None
-
-
-class Stats(StatsBase):
-    id: int
-
-
-class LoginRequest(BaseModel):
-    email: EmailStr
-    password: str
-
-
-class TokenResponse(BaseModel):
-    access_token: str
-    token_type: str = "bearer"
-
-
-roles_store: Dict[int, Role] = {}
-users_store: Dict[int, UserInDB] = {}
-alerts_store: Dict[int, Alert] = {}
-categories_store: Dict[int, Category] = {}
-notifications_store: Dict[int, Notification] = {}
-information_sources_store: Dict[int, InformationSource] = {}
-rss_channels_store: Dict[int, RSSChannel] = {}
-stats_store: Dict[int, Stats] = {}
-
-active_tokens: Dict[str, int] = {}
-
-counters = {
-    "roles": 1,
-    "users": 1,
-    "alerts": 1,
-    "categories": 1,
-    "notifications": 1,
-    "information_sources": 1,
-    "rss_channels": 1,
-    "stats": 1,
-}
-
-
-def next_id(counter_key: str) -> int:
-    value = counters[counter_key]
-    counters[counter_key] += 1
-    return value
-
-
-def ensure_role_ids_exist(role_ids: List[int]) -> None:
-    missing = [role_id for role_id in role_ids if role_id not in roles_store]
-    if missing:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Roles no encontrados: {missing}",
-        )
-
-
-def ensure_user_exists(user_id: int) -> None:
-    if user_id not in users_store:
-        raise HTTPException(status_code=404, detail="Usuario no encontrado")
-
-
-def ensure_alert_for_user(user_id: int, alert_id: int) -> Alert:
-    alert = alerts_store.get(alert_id)
-    if not alert or alert.user_id != user_id:
-        raise HTTPException(status_code=404, detail="Alerta no encontrada para el usuario")
-    return alert
-
-
-def ensure_notification_for_alert(alert_id: int, notification_id: int) -> Notification:
-    notification = notifications_store.get(notification_id)
-    if not notification or notification.alert_id != alert_id:
-        raise HTTPException(status_code=404, detail="Notificación no encontrada para la alerta")
-    return notification
-
-
-def ensure_information_source_exists(source_id: int) -> None:
-    if source_id not in information_sources_store:
-        raise HTTPException(status_code=404, detail="Fuente de información no encontrada")
-
-
-def ensure_category_exists(category_id: int) -> None:
-    if category_id not in categories_store:
-        raise HTTPException(status_code=404, detail="Categoría no encontrada")
-
-
-def ensure_rss_for_source(source_id: int, channel_id: int) -> RSSChannel:
-    channel = rss_channels_store.get(channel_id)
-    if not channel or channel.information_source_id != source_id:
-        raise HTTPException(status_code=404, detail="Canal RSS no encontrado para la fuente")
-    return channel
-
-
-def sanitize_user(user: UserInDB) -> User:
-    return User(
-        id=user.id,
-        email=user.email,
-        first_name=user.first_name,
-        last_name=user.last_name,
-        organization=user.organization,
-        role_ids=user.role_ids,
-    )
-
-
-def get_current_user(
-    credentials: HTTPAuthorizationCredentials = Depends(security),
-) -> UserInDB:
-    if credentials is None or credentials.scheme.lower() != "bearer":
-        raise HTTPException(status_code=401, detail="Token inválido o ausente")
-
-    user_id = active_tokens.get(credentials.credentials)
-    if not user_id:
-        raise HTTPException(status_code=401, detail="Token inválido o expirado")
-
-    user = users_store.get(user_id)
-    if not user:
-        raise HTTPException(status_code=401, detail="Usuario inválido")
-
-    return user
-
-
-def create_seed_data() -> None:
-    if roles_store:
-        return
-
-    admin_role_id = next_id("roles")
-    roles_store[admin_role_id] = Role(id=admin_role_id, name="admin")
-
-    user_role_id = next_id("roles")
-    roles_store[user_role_id] = Role(id=user_role_id, name="user")
-
-    admin_user_id = next_id("users")
-    users_store[admin_user_id] = UserInDB(
-        id=admin_user_id,
-        email="admin@newsradar.com",
-        first_name="Admin",
-        last_name="NewsRadar",
-        organization="NewsRadar",
-        role_ids=[admin_role_id],
-        password="admin123",
-    )
-
-
-@app.on_event("startup")
-def on_startup() -> None:
-    create_seed_data()
-
-
-@app.get(f"{API_PREFIX}/health", tags=["system"])
-def health() -> dict:
-    return {"status": "ok", "timestamp": datetime.now(timezone.utc).isoformat()}
-
-
-@app.post(f"{API_PREFIX}/auth/login", response_model=TokenResponse, tags=["auth"])
-def login(payload: LoginRequest) -> TokenResponse:
-    user = next((u for u in users_store.values() if u.email == payload.email), None)
-    if user is None or user.password != payload.password:
-        raise HTTPException(status_code=401, detail="Credenciales inválidas")
-
-    token = str(uuid4())
-    active_tokens[token] = user.id
-    return TokenResponse(access_token=token)
-
-
-@app.post(f"{API_PREFIX}/auth/register", response_model=User, tags=["auth"])
-def register(payload: UserCreate) -> User:
-    if any(user.email == payload.email for user in users_store.values()):
-        raise HTTPException(status_code=409, detail="El email ya está registrado")
-
-    ensure_role_ids_exist(payload.role_ids)
-
-    user_id = next_id("users")
-    user_db = UserInDB(id=user_id, **payload.model_dump())
-    users_store[user_id] = user_db
-    return sanitize_user(user_db)
-
-
-@app.get(f"{API_PREFIX}/users", response_model=List[User], tags=["users"])
-def list_users(_: UserInDB = Depends(get_current_user)) -> List[User]:
-    return [sanitize_user(user) for user in users_store.values()]
-
-
-@app.post(f"{API_PREFIX}/users", response_model=User, status_code=201, tags=["users"])
-def create_user(payload: UserCreate, _: UserInDB = Depends(get_current_user)) -> User:
-    if any(user.email == payload.email for user in users_store.values()):
-        raise HTTPException(status_code=409, detail="El email ya está registrado")
-
-    ensure_role_ids_exist(payload.role_ids)
-    user_id = next_id("users")
-    user_db = UserInDB(id=user_id, **payload.model_dump())
-    users_store[user_id] = user_db
-    return sanitize_user(user_db)
-
-
-@app.get(f"{API_PREFIX}/users/{{user_id}}", response_model=User, tags=["users"])
-def get_user(user_id: int, _: UserInDB = Depends(get_current_user)) -> User:
-    user = users_store.get(user_id)
-    if not user:
-        raise HTTPException(status_code=404, detail="Usuario no encontrado")
-    return sanitize_user(user)
-
-
-@app.put(f"{API_PREFIX}/users/{{user_id}}", response_model=User, tags=["users"])
-def update_user(user_id: int, payload: UserUpdate, _: UserInDB = Depends(get_current_user)) -> User:
-    user = users_store.get(user_id)
-    if not user:
-        raise HTTPException(status_code=404, detail="Usuario no encontrado")
-
-    data = payload.model_dump(exclude_unset=True)
-    if "email" in data and any(u.email == data["email"] and u.id != user_id for u in users_store.values()):
-        raise HTTPException(status_code=409, detail="El email ya está registrado")
-    if "role_ids" in data:
-        ensure_role_ids_exist(data["role_ids"])
-
-    updated = user.model_copy(update=data)
-    users_store[user_id] = updated
-    return sanitize_user(updated)
-
-
-@app.delete(
-    f"{API_PREFIX}/users/{{user_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["users"],
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
 )
-def delete_user(user_id: int, _: UserInDB = Depends(get_current_user)) -> None:
-    if user_id not in users_store:
-        raise HTTPException(status_code=404, detail="Usuario no encontrado")
 
-    alert_ids = [alert.id for alert in alerts_store.values() if alert.user_id == user_id]
-    for alert_id in alert_ids:
-        notification_ids = [n.id for n in notifications_store.values() if n.alert_id == alert_id]
-        for notification_id in notification_ids:
-            notifications_store.pop(notification_id, None)
-        alerts_store.pop(alert_id, None)
+app.include_router(api_router)
 
-    users_store.pop(user_id, None)
-
-
-@app.get(f"{API_PREFIX}/roles", response_model=List[Role], tags=["roles"])
-def list_roles(_: UserInDB = Depends(get_current_user)) -> List[Role]:
-    return list(roles_store.values())
-
-
-@app.post(f"{API_PREFIX}/roles", response_model=Role, status_code=201, tags=["roles"])
-def create_role(payload: RoleCreate, _: UserInDB = Depends(get_current_user)) -> Role:
-    role_id = next_id("roles")
-    role = Role(id=role_id, **payload.model_dump())
-    roles_store[role_id] = role
-    return role
-
-
-@app.get(f"{API_PREFIX}/roles/{{role_id}}", response_model=Role, tags=["roles"])
-def get_role(role_id: int, _: UserInDB = Depends(get_current_user)) -> Role:
-    role = roles_store.get(role_id)
-    if not role:
-        raise HTTPException(status_code=404, detail="Rol no encontrado")
-    return role
-
-
-@app.put(f"{API_PREFIX}/roles/{{role_id}}", response_model=Role, tags=["roles"])
-def update_role(role_id: int, payload: RoleUpdate, _: UserInDB = Depends(get_current_user)) -> Role:
-    role = roles_store.get(role_id)
-    if not role:
-        raise HTTPException(status_code=404, detail="Rol no encontrado")
-    updated = role.model_copy(update=payload.model_dump(exclude_unset=True))
-    roles_store[role_id] = updated
-    return updated
-
-
-@app.delete(
-    f"{API_PREFIX}/roles/{{role_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["roles"],
-)
-def delete_role(role_id: int, _: UserInDB = Depends(get_current_user)) -> None:
-    if role_id not in roles_store:
-        raise HTTPException(status_code=404, detail="Rol no encontrado")
-
-    for user in users_store.values():
-        if role_id in user.role_ids:
-            raise HTTPException(
-                status_code=409,
-                detail="No se puede eliminar un rol asignado a usuarios",
-            )
-
-    roles_store.pop(role_id, None)
-
-
-@app.get(
-    f"{API_PREFIX}/users/{{user_id}}/alerts",
-    response_model=List[Alert],
-    tags=["alerts"],
-)
-def list_user_alerts(user_id: int, _: UserInDB = Depends(get_current_user)) -> List[Alert]:
-    ensure_user_exists(user_id)
-    return [alert for alert in alerts_store.values() if alert.user_id == user_id]
-
-
-@app.post(
-    f"{API_PREFIX}/users/{{user_id}}/alerts",
-    response_model=Alert,
-    status_code=201,
-    tags=["alerts"],
-)
-def create_user_alert(user_id: int, payload: AlertCreate, _: UserInDB = Depends(get_current_user)) -> Alert:
-    ensure_user_exists(user_id)
-    alert_id = next_id("alerts")
-    alert = Alert(id=alert_id, user_id=user_id, **payload.model_dump())
-    alerts_store[alert_id] = alert
-    return alert
-
-
-@app.get(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}",
-    response_model=Alert,
-    tags=["alerts"],
-)
-def get_user_alert(user_id: int, alert_id: int, _: UserInDB = Depends(get_current_user)) -> Alert:
-    return ensure_alert_for_user(user_id, alert_id)
-
-
-@app.put(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}",
-    response_model=Alert,
-    tags=["alerts"],
-)
-def update_user_alert(
-    user_id: int,
-    alert_id: int,
-    payload: AlertUpdate,
-    _: UserInDB = Depends(get_current_user),
-) -> Alert:
-    alert = ensure_alert_for_user(user_id, alert_id)
-    updated = alert.model_copy(update=payload.model_dump(exclude_unset=True))
-    alerts_store[alert_id] = updated
-    return updated
-
-
-@app.delete(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["alerts"],
-)
-def delete_user_alert(user_id: int, alert_id: int, _: UserInDB = Depends(get_current_user)) -> None:
-    ensure_alert_for_user(user_id, alert_id)
-    notification_ids = [n.id for n in notifications_store.values() if n.alert_id == alert_id]
-    for notification_id in notification_ids:
-        notifications_store.pop(notification_id, None)
-    alerts_store.pop(alert_id, None)
-
-
-@app.get(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}/notifications",
-    response_model=List[Notification],
-    tags=["notifications"],
-)
-def list_alert_notifications(
-    user_id: int,
-    alert_id: int,
-    _: UserInDB = Depends(get_current_user),
-) -> List[Notification]:
-    ensure_alert_for_user(user_id, alert_id)
-    return [item for item in notifications_store.values() if item.alert_id == alert_id]
-
-
-@app.post(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}/notifications",
-    response_model=Notification,
-    status_code=201,
-    tags=["notifications"],
-)
-def create_alert_notification(
-    user_id: int,
-    alert_id: int,
-    payload: NotificationCreate,
-    _: UserInDB = Depends(get_current_user),
-) -> Notification:
-    ensure_alert_for_user(user_id, alert_id)
-    notification_id = next_id("notifications")
-    notification = Notification(id=notification_id, alert_id=alert_id, **payload.model_dump())
-    notifications_store[notification_id] = notification
-    return notification
-
-
-@app.get(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}/notifications/{{notification_id}}",
-    response_model=Notification,
-    tags=["notifications"],
-)
-def get_alert_notification(
-    user_id: int,
-    alert_id: int,
-    notification_id: int,
-    _: UserInDB = Depends(get_current_user),
-) -> Notification:
-    ensure_alert_for_user(user_id, alert_id)
-    return ensure_notification_for_alert(alert_id, notification_id)
-
-
-@app.put(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}/notifications/{{notification_id}}",
-    response_model=Notification,
-    tags=["notifications"],
-)
-def update_alert_notification(
-    user_id: int,
-    alert_id: int,
-    notification_id: int,
-    payload: NotificationUpdate,
-    _: UserInDB = Depends(get_current_user),
-) -> Notification:
-    ensure_alert_for_user(user_id, alert_id)
-    notification = ensure_notification_for_alert(alert_id, notification_id)
-    updated = notification.model_copy(update=payload.model_dump(exclude_unset=True))
-    notifications_store[notification_id] = updated
-    return updated
-
-
-@app.delete(
-    f"{API_PREFIX}/users/{{user_id}}/alerts/{{alert_id}}/notifications/{{notification_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["notifications"],
-)
-def delete_alert_notification(
-    user_id: int,
-    alert_id: int,
-    notification_id: int,
-    _: UserInDB = Depends(get_current_user),
-) -> None:
-    ensure_alert_for_user(user_id, alert_id)
-    ensure_notification_for_alert(alert_id, notification_id)
-    notifications_store.pop(notification_id, None)
-
-
-@app.get(f"{API_PREFIX}/categories", response_model=List[Category], tags=["categories"])
-def list_categories(_: UserInDB = Depends(get_current_user)) -> List[Category]:
-    return list(categories_store.values())
-
-
-@app.post(f"{API_PREFIX}/categories", response_model=Category, status_code=201, tags=["categories"])
-def create_category(payload: CategoryCreate, _: UserInDB = Depends(get_current_user)) -> Category:
-    category_id = next_id("categories")
-    category = Category(id=category_id, **payload.model_dump())
-    categories_store[category_id] = category
-    return category
-
-
-@app.get(f"{API_PREFIX}/categories/{{category_id}}", response_model=Category, tags=["categories"])
-def get_category(category_id: int, _: UserInDB = Depends(get_current_user)) -> Category:
-    category = categories_store.get(category_id)
-    if not category:
-        raise HTTPException(status_code=404, detail="Categoría no encontrada")
-    return category
-
-
-@app.put(f"{API_PREFIX}/categories/{{category_id}}", response_model=Category, tags=["categories"])
-def update_category(category_id: int, payload: CategoryUpdate, _: UserInDB = Depends(get_current_user)) -> Category:
-    category = categories_store.get(category_id)
-    if not category:
-        raise HTTPException(status_code=404, detail="Categoría no encontrada")
-    updated = category.model_copy(update=payload.model_dump(exclude_unset=True))
-    categories_store[category_id] = updated
-    return updated
-
-
-@app.delete(
-    f"{API_PREFIX}/categories/{{category_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["categories"],
-)
-def delete_category(category_id: int, _: UserInDB = Depends(get_current_user)) -> None:
-    if category_id not in categories_store:
-        raise HTTPException(status_code=404, detail="Categoría no encontrada")
-
-    for channel in rss_channels_store.values():
-        if channel.category_id == category_id:
-            raise HTTPException(status_code=409, detail="Categoría asociada a canales RSS")
-
-    categories_store.pop(category_id, None)
-
-
-@app.get(
-    f"{API_PREFIX}/information-sources",
-    response_model=List[InformationSource],
-    tags=["information-sources"],
-)
-def list_information_sources(_: UserInDB = Depends(get_current_user)) -> List[InformationSource]:
-    return list(information_sources_store.values())
-
-
-@app.post(
-    f"{API_PREFIX}/information-sources",
-    response_model=InformationSource,
-    status_code=201,
-    tags=["information-sources"],
-)
-def create_information_source(
-    payload: InformationSourceCreate,
-    _: UserInDB = Depends(get_current_user),
-) -> InformationSource:
-    source_id = next_id("information_sources")
-    source = InformationSource(id=source_id, **payload.model_dump())
-    information_sources_store[source_id] = source
-    return source
-
-
-@app.get(
-    f"{API_PREFIX}/information-sources/{{source_id}}",
-    response_model=InformationSource,
-    tags=["information-sources"],
-)
-def get_information_source(source_id: int, _: UserInDB = Depends(get_current_user)) -> InformationSource:
-    source = information_sources_store.get(source_id)
-    if not source:
-        raise HTTPException(status_code=404, detail="Fuente de información no encontrada")
-    return source
-
-
-@app.put(
-    f"{API_PREFIX}/information-sources/{{source_id}}",
-    response_model=InformationSource,
-    tags=["information-sources"],
-)
-def update_information_source(
-    source_id: int,
-    payload: InformationSourceUpdate,
-    _: UserInDB = Depends(get_current_user),
-) -> InformationSource:
-    source = information_sources_store.get(source_id)
-    if not source:
-        raise HTTPException(status_code=404, detail="Fuente de información no encontrada")
-    updated = source.model_copy(update=payload.model_dump(exclude_unset=True))
-    information_sources_store[source_id] = updated
-    return updated
-
-
-@app.delete(
-    f"{API_PREFIX}/information-sources/{{source_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["information-sources"],
-)
-def delete_information_source(source_id: int, _: UserInDB = Depends(get_current_user)) -> None:
-    if source_id not in information_sources_store:
-        raise HTTPException(status_code=404, detail="Fuente de información no encontrada")
-
-    channel_ids = [
-        channel.id
-        for channel in rss_channels_store.values()
-        if channel.information_source_id == source_id
-    ]
-    for channel_id in channel_ids:
-        rss_channels_store.pop(channel_id, None)
-
-    information_sources_store.pop(source_id, None)
-
-
-@app.get(
-    f"{API_PREFIX}/information-sources/{{source_id}}/rss-channels",
-    response_model=List[RSSChannel],
-    tags=["rss-channels"],
-)
-def list_source_channels(source_id: int, _: UserInDB = Depends(get_current_user)) -> List[RSSChannel]:
-    ensure_information_source_exists(source_id)
-    return [
-        channel
-        for channel in rss_channels_store.values()
-        if channel.information_source_id == source_id
-    ]
-
-
-@app.post(
-    f"{API_PREFIX}/information-sources/{{source_id}}/rss-channels",
-    response_model=RSSChannel,
-    status_code=201,
-    tags=["rss-channels"],
-)
-def create_source_channel(
-    source_id: int,
-    payload: RSSChannelCreate,
-    _: UserInDB = Depends(get_current_user),
-) -> RSSChannel:
-    ensure_information_source_exists(source_id)
-    ensure_category_exists(payload.category_id)
-
-    channel_id = next_id("rss_channels")
-    channel = RSSChannel(
-        id=channel_id,
-        information_source_id=source_id,
-        **payload.model_dump(),
-    )
-    rss_channels_store[channel_id] = channel
-    return channel
-
-
-@app.get(
-    f"{API_PREFIX}/information-sources/{{source_id}}/rss-channels/{{channel_id}}",
-    response_model=RSSChannel,
-    tags=["rss-channels"],
-)
-def get_source_channel(
-    source_id: int,
-    channel_id: int,
-    _: UserInDB = Depends(get_current_user),
-) -> RSSChannel:
-    ensure_information_source_exists(source_id)
-    return ensure_rss_for_source(source_id, channel_id)
-
-
-@app.put(
-    f"{API_PREFIX}/information-sources/{{source_id}}/rss-channels/{{channel_id}}",
-    response_model=RSSChannel,
-    tags=["rss-channels"],
-)
-def update_source_channel(
-    source_id: int,
-    channel_id: int,
-    payload: RSSChannelUpdate,
-    _: UserInDB = Depends(get_current_user),
-) -> RSSChannel:
-    ensure_information_source_exists(source_id)
-    channel = ensure_rss_for_source(source_id, channel_id)
-
-    update_data = payload.model_dump(exclude_unset=True)
-    if "category_id" in update_data:
-        ensure_category_exists(update_data["category_id"])
-
-    updated = channel.model_copy(update=update_data)
-    rss_channels_store[channel_id] = updated
-    return updated
-
-
-@app.delete(
-    f"{API_PREFIX}/information-sources/{{source_id}}/rss-channels/{{channel_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["rss-channels"],
-)
-def delete_source_channel(
-    source_id: int,
-    channel_id: int,
-    _: UserInDB = Depends(get_current_user),
-) -> None:
-    ensure_information_source_exists(source_id)
-    ensure_rss_for_source(source_id, channel_id)
-    rss_channels_store.pop(channel_id, None)
-
-
-@app.get(f"{API_PREFIX}/stats", response_model=List[Stats], tags=["stats"])
-def list_stats(_: UserInDB = Depends(get_current_user)) -> List[Stats]:
-    return list(stats_store.values())
-
-
-@app.post(f"{API_PREFIX}/stats", response_model=Stats, status_code=201, tags=["stats"])
-def create_stats(payload: StatsCreate, _: UserInDB = Depends(get_current_user)) -> Stats:
-    stats_id = next_id("stats")
-    stats = Stats(id=stats_id, **payload.model_dump())
-    stats_store[stats_id] = stats
-    return stats
-
-
-@app.get(f"{API_PREFIX}/stats/{{stats_id}}", response_model=Stats, tags=["stats"])
-def get_stats(stats_id: int, _: UserInDB = Depends(get_current_user)) -> Stats:
-    stats = stats_store.get(stats_id)
-    if not stats:
-        raise HTTPException(status_code=404, detail="Stats no encontrados")
-    return stats
-
-
-@app.put(f"{API_PREFIX}/stats/{{stats_id}}", response_model=Stats, tags=["stats"])
-def update_stats(stats_id: int, payload: StatsUpdate, _: UserInDB = Depends(get_current_user)) -> Stats:
-    stats = stats_store.get(stats_id)
-    if not stats:
-        raise HTTPException(status_code=404, detail="Stats no encontrados")
-
-    updated = stats.model_copy(update=payload.model_dump(exclude_unset=True))
-    stats_store[stats_id] = updated
-    return updated
-
-
-@app.delete(
-    f"{API_PREFIX}/stats/{{stats_id}}",
-    status_code=204,
-    response_model=None,
-    response_class=Response,
-    tags=["stats"],
-)
-def delete_stats(stats_id: int, _: UserInDB = Depends(get_current_user)) -> None:
-    if stats_id not in stats_store:
-        raise HTTPException(status_code=404, detail="Stats no encontrados")
-    stats_store.pop(stats_id, None)
+# Añadimos un endpoint raíz para saber que el motor funciona
+@app.get("/", tags=["system"])
+def root():
+    return {"message": "Motor API REST de NewsRadar activo. Visita /docs"}
